feat: seperate share-pool-modal from pool list page for admin
